edit controller task

+++ ME 507 Clawmachine/src/motor1_task.cpp
@@ -33,14 +33,10 @@ void task_move_motor (void* p_params)
     uint16_t currentencoderx;
     share_encoder_positionx.get(currentencoderx);
 
-    
-    
-
     if (usergivenx != currentencoderx) // if the current encoder position isnt equal to the user desired than do the following
     //when making other motor task use shares for y since that is for the other encoder
     {
      object1.setduty(35); // keep moving
-
     }
 // error below
     else 
@@ -54,29 +50,17 @@ void task_move_motor (void* p_params)
       else
         {
          object1.setduty(0); 
-         uint8_t jobstatus = 1; // job is complete
-         
+         uint16_t jobstatus = 1; // job is complete
          share_job_status.put(jobstatus); // put number 1 into share variable to send to main
-
          }
-     
       //  it checks if its returned to its orgin
-
     // then sends the share bool that the job is completed 
      }
-
  }
-
-    
-    // set up code goes here
-// check encoder val, if correct spot stop spinning otherwise continue setduty funct
-// set up shares to get encoder values 
 
     for (;;)
     {
-        delay (100);
-
-        
+        delay (100);   
     }
 }
